+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 ### Fixed
 - A lot of code style warnings (e.g. legacy java 6 code can be updated, etc)
+### Added
+- TestItemTree.ItemTreeKey now supports attribute storing
 
 ## [5.0.12]
 ### Fixed